refactor: extract pagination hook and component
