+++ frontend/utils/string.ts
@@ -1,3 +1,7 @@
+export function isNumeric(value: string | null) {
+  return value !== null && /^-?\d+$/.test(value);
+}
+
 export function getNameFromEmail(email?: string) {
   return email?.split("@")[0];
 }
